Moved authContext and reducer to independant files.
