+++ src/tests/routes/protectedRoute.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { AuthenticatedRoute } from 'routes/protectedRoute';
 import Home from 'containers/Home';
-import { AuthProvider } from 'contexts/authContext';
+import { AuthProvider } from 'contexts/auth';
 import { Router } from 'react-router-dom';
 import { createMemoryHistory } from 'history';
